mlops_pipeline/requirements.txt
tensorflow
mlflow
fastapi
uvicorn
pillow
